test: TestConfigUpdate - new config
